+++ tests/sdk/dotnet/helm-api-versions/step1/Program.cs
@@ -1,10 +1,11 @@
-﻿// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+﻿// Copyright 2016-2022, Pulumi Corporation.  All rights reserved.
 
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi;
+using Pulumi.Kubernetes;
 using Pulumi.Kubernetes.Core.V1;
 using Pulumi.Kubernetes.Helm;
 using Pulumi.Kubernetes.Helm.V2;
@@ -16,7 +17,8 @@ class HelmStack : Stack
 {
     public HelmStack()
     {
-        var namespaceTest = new Namespace("test");
+        var provider = new Provider("k8s");
+        var namespaceTest = new Namespace("test", null, new CustomResourceOptions{Provider = provider});
         var namespaceName = namespaceTest.Metadata.Apply(n => n.Name);
       
         new Chart("api-versions", new LocalChartArgs
@@ -24,6 +26,9 @@ class HelmStack : Stack
             ApiVersions = { "foo", "bar" },
             Namespace = namespaceName,
             Path = "helm-api-versions"
+        }, new ComponentResourceOptions
+        {
+            Provider = provider,
         });
 
         new Chart("single-api-version", new LocalChartArgs
@@ -31,6 +36,9 @@ class HelmStack : Stack
             ApiVersions = { "foo" },
             Namespace = namespaceName,
             Path = "helm-single-api-version"
+        }, new ComponentResourceOptions
+        {
+            Provider = provider,
         });
     }
 }