+++ tests/sdk/dotnet/yaml-url/Program.cs
@@ -1,9 +1,10 @@
-﻿// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+﻿// Copyright 2016-2022, Pulumi Corporation.  All rights reserved.
 
 using System;
 using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi;
+using Pulumi.Kubernetes;
 using Pulumi.Kubernetes.Core.V1;
 using Pulumi.Kubernetes.Types.Inputs.Core.V1;
 using Pulumi.Kubernetes.Yaml;
@@ -12,16 +13,18 @@ class YamlStack : Stack
 {
     public YamlStack()
     {
+        var provider = new Provider("k8s");
+
         // Create two test namespaces to allow test parallelism.
-        var namespace1 = new Namespace("test-namespace", new NamespaceArgs());
-        var namespace2 = new Namespace("test-namespace2", new NamespaceArgs());
+        var namespace1 = new Namespace("test-namespace", new NamespaceArgs(), new CustomResourceOptions{Provider = provider});
+        var namespace2 = new Namespace("test-namespace2", new NamespaceArgs(), new CustomResourceOptions{Provider = provider});
 
         // Create resources from standard Kubernetes guestbook YAML example in the first test namespace.
-        var file1 = namespace1.Metadata.Apply(m => m.Name).Apply(ns => MakeConfigFile("guestbook", ns));
+        var file1 = namespace1.Metadata.Apply(m => m.Name).Apply(ns => MakeConfigFile("guestbook", ns, provider));
 
         // Create resources from standard Kubernetes guestbook YAML example in the second test namespace.
         // Disambiguate resource names with a specified prefix.
-        var file2 = namespace2.Metadata.Apply(m => m.Name).Apply(ns => MakeConfigFile("guestbook", ns, "dup"));
+        var file2 = namespace2.Metadata.Apply(m => m.Name).Apply(ns => MakeConfigFile("guestbook", ns, provider, "dup"));
 
         this.FileUrns = Output.Format($"{file1.Apply(f => f.Urn)},{file2.Apply(f => f.Urn)}");
     }
@@ -29,7 +32,7 @@ class YamlStack : Stack
     [Output]
     public Output<string> FileUrns { get; set; }
 
-    private ConfigFile MakeConfigFile(string name, string namespaceName, string? resourcePrefix = null)
+    private ConfigFile MakeConfigFile(string name, string namespaceName, Provider provider, string? resourcePrefix = null)
     {
         return new ConfigFile(name, new ConfigFileArgs
         {
@@ -46,6 +49,9 @@ class YamlStack : Stack
                     return result.SetItem("metadata", newMeta);
                 }
             }
+        }, new ComponentResourceOptions
+        {
+            Provider = provider,
         });
     }
 }